Created a new listener for conflicts tree
